+++ libs/utils/src/index.ts
@@ -4,5 +4,6 @@ export * from './lib/apply-attrs/apply-attrs.directive';
 export * from './lib/of-type.pipe';
 export * from './lib/of-type.module';
 export * from './lib/to-json';
+export * from './lib/to-boolean';
 export * from './lib/apply-contexts';
 export * from './lib/context';