+++ libs/select/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-next.4](https://github.com/spryker/ui-components/compare/@spryker/select@0.1.0-next.3...@spryker/select@0.1.0-next.4) (2020-03-12)
+
+**Note:** Version bump only for package @spryker/select
+
+
+
+
+
 # [0.1.0-next.3](https://github.com/spryker/ui-components/compare/@spryker/select@0.1.0-next.2...@spryker/select@0.1.0-next.3) (2020-03-12)